refactor: 공지 서비스 패키지 변경
