+++ src/main/java/shop/tryit/domain/notice/service/NoticeService.java
@@ -1,4 +1,4 @@
-package shop.tryit.domain.notice;
+package shop.tryit.domain.notice.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -7,6 +7,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import shop.tryit.domain.notice.NoticeRepository;
 import shop.tryit.domain.notice.entity.Notice;
 
 @Service
