+++ src/test/java/shop/tryit/domain/notice/NoticeServiceTests.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 import shop.tryit.domain.notice.entity.Notice;
+import shop.tryit.domain.notice.service.NoticeService;
 
 @Transactional
 @SpringBootTest
